# openapi7.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hamdaz ERP",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/analytics/preview": {
   "get": {
    "tags": [
     "user analytics"
    ],
    "summary": "Rank people for the next assignment, keeping nothing",
    "description": "Who should get the next task, and why.\n\nA team is required, and it must have an assignment policy of its own \u2014 only\nthose distribute work through this. There is deliberately no\norganisation-wide ranking: scoring everybody at once would put every team in\nthe company into one queue they do not share.\n\nTask counts are read live. Nothing is stored, in this database or anywhere\nelse \u2014 run it as often as you like.",
    "operationId": "preview_api_v1_analytics_preview_get",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Team handle. Required.",
       "title": "Team"
      },
      "description": "Team handle. Required."
     },
     {
      "name": "refresh",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Re-read SharePoint instead of the cache",
       "default": false,
       "title": "Refresh"
      },
      "description": "Re-read SharePoint instead of the cache"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analytics/runs": {
   "post": {
    "tags": [
     "user analytics"
    ],
    "summary": "Compute a ranking and keep it",
    "description": "Store the ranking as the record of a decision.\n\nKept in Postgres. The policy is frozen onto the row, so editing it next week\ncannot change what this run appears to have been based on.\n\nRestricted to the people who may edit the policy in the first place \u2014 a\nstored run is a record others will rely on, so it should not accumulate from\nanyone who happens to look.",
    "operationId": "create_run_api_v1_analytics_runs_post",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Team handle. Required.",
       "title": "Team"
      },
      "description": "Team handle. Required."
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "user analytics"
    ],
    "summary": "Kept runs",
    "operationId": "index_api_v1_analytics_runs_get",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only this team's runs",
       "title": "Team"
      },
      "description": "Only this team's runs"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RunSummaryOut"
         },
         "title": "Response Index Api V1 Analytics Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analytics/runs/{run_id}": {
   "get": {
    "tags": [
     "user analytics"
    ],
    "summary": "One kept run in full",
    "operationId": "detail_api_v1_analytics_runs__run_id__get",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "user analytics"
    ],
    "summary": "Delete a kept run",
    "operationId": "remove_api_v1_analytics_runs__run_id__delete",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analytics/runs/{run_id}/people/{user_id}": {
   "get": {
    "tags": [
     "user analytics"
    ],
    "summary": "Why one person scored what they did",
    "description": "The factor breakdown for one person.\n\nThis is the endpoint that makes a ranking arguable: it shows the raw number,\nwhere it sat relative to everyone else, and how much each factor moved them.",
    "operationId": "explain_api_v1_analytics_runs__run_id__people__user_id__get",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Run Id"
      }
     },
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EntryOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EntryOut": {
    "properties": {
     "user_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "sharepoint_lookup_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sharepoint Lookup Id"
     },
     "total_tasks": {
      "type": "integer",
      "title": "Total Tasks"
     },
     "open_tasks": {
      "type": "integer",
      "title": "Open Tasks"
     },
     "completed_tasks": {
      "type": "integer",
      "title": "Completed Tasks"
     },
     "overdue_tasks": {
      "type": "integer",
      "title": "Overdue Tasks"
     },
     "due_soon_tasks": {
      "type": "integer",
      "title": "Due Soon Tasks"
     },
     "no_status_tasks": {
      "type": "integer",
      "title": "No Status Tasks"
     },
     "expired_tasks": {
      "type": "integer",
      "title": "Expired Tasks"
     },
     "bid_closed_tasks": {
      "type": "integer",
      "title": "Bid Closed Tasks"
     },
     "active_tasks": {
      "type": "integer",
      "title": "Active Tasks"
     },
     "last_assigned_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Assigned On"
     },
     "days_since_last_assign": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Days Since Last Assign"
     },
     "labels": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Labels"
     },
     "capacity": {
      "type": "string",
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Capacity"
     },
     "effective_load": {
      "type": "string",
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Effective Load"
     },
     "max_open": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Open"
     },
     "excluded": {
      "type": "boolean",
      "title": "Excluded"
     },
     "excluded_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Excluded Reason"
     },
     "priority_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Priority Score"
     },
     "factor_total": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Factor Total"
     },
     "factors": {
      "additionalProperties": {
       "$ref": "#/components/schemas/FactorOut"
      },
      "type": "object",
      "title": "Factors"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "display_name",
     "email",
     "sharepoint_lookup_id",
     "total_tasks",
     "open_tasks",
     "completed_tasks",
     "overdue_tasks",
     "due_soon_tasks",
     "no_status_tasks",
     "expired_tasks",
     "bid_closed_tasks",
     "active_tasks",
     "last_assigned_on",
     "days_since_last_assign",
     "labels",
     "capacity",
     "effective_load",
     "max_open",
     "excluded",
     "excluded_reason",
     "priority_score",
     "factor_total",
     "factors"
    ],
    "title": "EntryOut"
   },
...
   "FactorOut": {
    "properties": {
     "raw": {
      "type": "number",
      "title": "Raw"
     },
     "normalised": {
      "type": "number",
      "title": "Normalised"
     },
     "weight": {
      "type": "number",
      "title": "Weight"
     },
     "contribution": {
      "type": "number",
      "title": "Contribution"
     }
    },
    "type": "object",
    "required": [
     "raw",
     "normalised",
     "weight",
     "contribution"
    ],
    "title": "FactorOut",
    "description": "One factor's part in one person's score."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RunIn": {
    "properties": {
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "title": "RunIn",
    "description": "Keep this run as the record of a decision."
   },
   "RunOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "team_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team Id"
     },
     "team_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team Name"
     },
     "policy_snapshot": {
      "additionalProperties": true,
      "type": "object",
      "title": "Policy Snapshot"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "rows_read": {
      "type": "integer",
      "title": "Rows Read"
     },
     "excluded_note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Excluded Note"
     },
     "saved": {
      "type": "boolean",
      "title": "Saved"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "created_by_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By Name"
     },
     "entries": {
      "items": {
       "$ref": "#/components/schemas/EntryOut"
      },
      "type": "array",
      "title": "Entries"
     },
     "next_up": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Up"
     },
     "assignable": {
      "type": "integer",
      "title": "Assignable",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "id",
     "team_id",
     "team_name",
     "policy_snapshot",
     "source",
     "rows_read",
     "excluded_note",
     "saved",
     "notes",
     "created_at",
     "entries"
    ],
    "title": "RunOut"
   },
   "RunSummaryOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "team_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team Name"
     },
     "rows_read": {
      "type": "integer",
      "title": "Rows Read"
     },
     "people": {
      "type": "integer",
      "title": "People"
     },
     "assignable": {
      "type": "integer",
      "title": "Assignable"
     },
     "next_up": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Up"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "created_by_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "team_name",
     "rows_read",
     "people",
     "assignable",
     "next_up",
     "notes",
     "created_at",
     "created_by_name"
    ],
    "title": "RunSummaryOut",
    "description": "A history row \u2014 no entries, so a list of fifty runs stays small."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}